refactor: move sanitization functions to models package
- Extract SanitizeString and SanitizeRows from genai/scraper.go
- Create new internal/models/sanitize.go with sanitization logic
- Move sanitization tests to internal/models/sanitize_test.go
- Update genai/scraper.go to use models.SanitizeString and models.SanitizeRows
- Remove deprecated sanitization code from genai package

This centralizes sanitization logic in the models package where it
belongs, making it reusable across different modules.
